+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added JSDoc for functions
 - added warnings about pre-production usage and simple roadmap about versioning and release schedule
 - concat function
+- equals function
 
 ### Changed
 - added more types